omni-ai/requirements.txt
Flask
flask-cors
google-genai==0.8.0
Pillow
python-dotenv
google-generativeai
